refactor: move client creation to fetcher.go
